fix: bugbot complaints
